pyproject.toml
[tool.poetry]
name = "codeflow"
version = "0.1.0"
description = ""
authors = ["Erik Nomitch <enomitch@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
llama-index = "^0.7.11.post1"
python-dotenv = "^1.0.0"
llama-hub = "^0.0.14"
langchain = "^0.0.240"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
